% Generated by roxygen2: do not edit by hand
% Please edit documentation in r/arithemtic_fun.R
\name{adder}
\alias{adder}
\title{Add two numbers together}
\usage{
adder(a, b)
}
\arguments{
\item{a}{Numeric vector; The first of two numbers to add}

\item{b}{Numeric vector}
}
\value{
Numeric vector of the same length as the input
}
\description{
Here is a description - has to be 1 paragraph long (but not really)
}
\details{
\code{a} and \code{b} can be vector of length grater than one, but should be the same length as each other. If the vectors are different length no recycling is carried, and an error is raised.
}
\examples{
adder(1, 2)
adder(1:3, 11:13)

}
