% Generated by roxygen2: do not edit by hand
% Please edit documentation in r/assign_usernames.R
\name{assign_usernames}
\alias{assign_usernames}
\title{assign_usernames}
\usage{
assign_usernames(names_vector)
}
\arguments{
\item{names_vector}{character vector of student names}
}
\value{
A dataframe of student names and usernames
}
\description{
A function that takes a vector of names and assigns usernames
to each for use in a server. Requires tidyverse.
}
\examples{
assign_usernames(c("John", "Colin"))
}
