% Generated by roxygen2: do not edit by hand
% Please edit documentation in r/does_al_approve.R
\name{does_al_approve}
\alias{does_al_approve}
\title{does_al_approve}
\usage{
does_al_approve(text)
}
\arguments{
\item{text}{character string}
}
\value{
A character vector
}
\description{
A function to add the words "Al approves this message" or
"Al doesn't approve this message" to a character input.
}
\examples{
does_al_approve("Pizza is the best food")
}
