% Generated by roxygen2: do not edit by hand
% Please edit documentation in r/check_readme_links.R
\name{find_broken_links}
\alias{find_broken_links}
\title{Find Broken Links}
\usage{
find_broken_links(dir = ".")
}
\arguments{
\item{dir}{Character vector showing local path to data rewrite 2 / classnotes folder}
}
\value{
Character vector of invalid paths
}
\description{
Finds broken links within a README.html file. Usually used to check that the Data Rewrite 2 Readme is up-to-date.
}
\examples{
find_broken_links('~/professional_data_analysis/course_notes/data_rewrite2')
}
