% Generated by roxygen2: do not edit by hand
% Please edit documentation in r/grouper.R
\name{grouper}
\alias{grouper}
\title{Create psuedo-random groups from a given vector of peoples' names}
\usage{
grouper(cohort, number_of_groups = 2)
}
\arguments{
\item{cohort}{Character vector}

\item{number_of_groups}{Integer value}
}
\description{
Create psuedo-random groups from a given vector of peoples' names
}
\examples{
grouper(c("al", "david", "jamie", "john", "neringa", "uzair"), 3)


}
