% Generated by roxygen2: do not edit by hand
% Please edit documentation in r/is_it_tidyverse.R
\name{is_it_tidyverse}
\alias{is_it_tidyverse}
\title{is_it_tidyverse}
\usage{
is_it_tidyverse(function_to_check)
}
\arguments{
\item{function_to_check}{character string}
}
\value{
A character vector
}
\description{
A function to check whether a function is part of the tidyverse packages
}
\examples{
is_it_tidyverse("select")
}
