% Generated by roxygen2: do not edit by hand
% Please edit documentation in r/arithemtic_fun.R
\name{product}
\alias{product}
\title{Title}
\usage{
product(a, b)
}
\arguments{
\item{a}{Numeric}

\item{b}{Numeric}
}
\description{
iterate ove fun til it works
using load_all() each time for any ad hoc test
when you happy do devtools::document()
iterate over docs til you're happy
}
